+++ src/card/sets/tests/runtime_support/stack_effects.rs
@@ -276,6 +276,7 @@ fn shared_stack_effect_at_position(effect: EffectDef, deferred_decision_allowed:
         | EffectDef::Detain { object }
         | EffectDef::GainControl { object, .. }
         | EffectDef::AddCounters { object, .. }
+        | EffectDef::RemoveCounters { object, .. }
         | EffectDef::Attach { object }
         | EffectDef::Reconfigure { object }
         | EffectDef::ChangeTextBasicLandType { object }